Setup navigation, change package structure
