+++ app/src/main/java/com/masslany/justevaluateit/presentation/dashboard/DashboardFragment.kt
@@ -1,0 +1,31 @@
+package com.masslany.justevaluateit.presentation.dashboard
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.platform.ComposeView
+import androidx.fragment.app.Fragment
+import com.masslany.justevaluateit.presentation.ui.theme.JustEvaluateItTheme
+
+class DashboardFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return ComposeView(requireContext()).apply {
+            setContent {
+                JustEvaluateItTheme {
+                    // A surface container using the 'background' color from the theme
+                    Surface(color = MaterialTheme.colors.background) {
+                        DashboardScreen()
+                    }
+                }
+            }
+        }
+    }
+}
