+++ app/src/main/java/com/masslany/justevaluateit/presentation/ui/theme/Dimentions.kt
@@ -1,4 +1,4 @@
-package com.masslany.justevaluateit.ui.theme
+package com.masslany.justevaluateit.presentation.ui.theme
 
 import androidx.compose.ui.unit.dp
